refactor: TagBadgeGroup 분리 및 콘텐츠 더보기 로직 단순화
