+++ apps/web/src/entities/feed/index.ts
@@ -1,2 +1,3 @@
 export { FEED_QUERY_KEY } from './model/queryKeys';
 export type { FeedPage, FeedPost, FeedTag } from './model/types';
+export { default as TagBadgeGroup } from './ui/TagBadgeGroup';
